NMR-6904: move FileProperty, ColorProperty and complex items in the nmrfx-utils
