+++ nmrfx-utils/src/main/java/org/nmrfx/utils/properties/ColorProperty.java
@@ -16,39 +16,31 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
- /*
- * To change this license header, choose License Headers in Project Properties.
- * To change this template file, choose Tools | Templates
- * and open the template in the editor.
- */
-package org.nmrfx.processor.gui;
+package org.nmrfx.utils.properties;
 
-import java.io.File;
 import javafx.beans.property.SimpleObjectProperty;
+import javafx.scene.paint.Color;
 
 /**
  *
  * @author brucejohnson
  */
-public class FileProperty extends SimpleObjectProperty<File> {
+public class ColorProperty extends SimpleObjectProperty<Color> {
 
-    File file = null;
+    public ColorProperty(Object object, String name, Color color) {
+        super(object, name, color);
+    }
 
-    @Override
-    public File get() {
-        return file;
+    public void setColor(Color color) {
+        set(color);
+    }
+
+    public Color getColor() {
+        return get();
     }
 
     @Override
-    public void setValue(File file) {
-        this.file = file;
-        this.fireValueChangedEvent();
+    public String toString() {
+        return super.toString() + " " + get();
     }
-
-    @Override
-    public void set(File file) {
-        this.file = file;
-        this.fireValueChangedEvent();
-    }
-
 }
